requirements.txt
numpy
opencv-python
pyautogui
kaggle
torch
torchvision
pillow
tqdm
matplotlib
